add functions & add :twitter-id
added a few missing clisk functions that could be interesting.

also, have code for issue #14 to add a twitter-id to the gist data.
Need to add an entry point, though.  Do this once per day via heroku run?
Twitter URL is
https://twitter.com/{screen-name}/status/{twitter-id}

finally, put in sketch of code (commented out) waiting for bump to
v1.8.0 for media-upload-chunked